requirements.txt
fastapi
uvicorn
python-dotenv
openai==1.12.0
pydantic
motor==3.3.2
pymongo==4.6.1
requests==2.31.0
uuid==1.30
pytest==7.4.3